+++ shared/src/model/config/panel_api.rs
@@ -1,0 +1,31 @@
+use crate::utils::{default_as_true};
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct PanelApiQueryParamDto {
+    pub key: String,
+    pub value: String,
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct PanelApiQueryParametersDto {
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub client_info: Vec<PanelApiQueryParamDto>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub client_new: Vec<PanelApiQueryParamDto>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub client_renew: Vec<PanelApiQueryParamDto>,
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct PanelApiConfigDto {
+    #[serde(default = "default_as_true")]
+    pub enabled: bool,
+    pub url: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub api_key: Option<String>,
+    #[serde(default)]
+    pub query_parameter: PanelApiQueryParametersDto,
+}